Move nvfp4_quant.py from gemm to common
