+++ modelopt/torch/kernels/quantization/attention/__init__.py
@@ -19,6 +19,6 @@
 ``@triton.jit`` helpers invoked by the unified flash-attention kernel in
 ``common/attention/triton_fa.py`` under its ``P_QDQ`` constexpr guard.
 Only NVFP4 needs a P-specific helper (tiling and block-amax policy on top of
-``quantization/gemm/nvfp4_quant.py``); the FP8 mode uses
+``quantization/common/nvfp4_quant.py``); the FP8 mode uses
 ``quantization/common/fp8_quant.fp8_scalar_qdq`` directly.
 """
